# docs/api/entitlements_openapi.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: DNS_HOST
basePath: /api/entitlements/v2
...
paths:
...
  '/members/{member_email}/groups':
    get:
      tags:
        - list-group-on-behalf-of-api
      summary: listGroupsOnBehalfOf
      operationId: listGroupsOnBehalfOfUsingGET
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: data-partition-id
          in: header
          description: tenant
          required: true
          type: string
          default: common
        - name: member_email
          in: path
          description: member_email
          required: true
          type: string
        - name: type
          in: query
          description: type
          required: true
          enum:
            - DATA
            - NONE
            - SERVICE
            - USER
        - name: appid
          in: query
          description: appid
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ListGroupResponseDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - JWT:
            - global
...
securityDefinitions:
  JWT:
    type: apiKey
    name: Authorization
    in: header
definitions:
...
  ListGroupResponseDto:
    type: object
...